+++ scripts/local_modules/painter.js
@@ -37,7 +37,7 @@ export default class Painter {
             draw: (typeof options.draw !== 'undefined') ? options.draw : true,
             color: (typeof options.color !== 'undefined') ? options.color : '#FFF',
             point: (typeof options.point !== 'undefined') ? options.point : 0.5,
-            backgroundFill: (typeof options.backgroundFill !== 'undefined') ? options.backgroundFill : '#050490'
+            backgroundFill: (typeof options.backgroundFill !== 'undefined') ? options.backgroundFill : '#1b5eec'
         };
         Object.assign(this, JSON.parse(JSON.stringify(this.settings)));
 
@@ -119,10 +119,10 @@ export default class Painter {
         let guides = this.guideContext;
         this.brushes[circle.id].forEach(brush => {
             const radians = circle.getRadians();
-            const x1 = circle.x0 + (Math.cos(radians) * circle.radius);
-            const y1 = circle.y0 + (Math.sin(radians) * circle.radius);
-            const x = x1 + (Math.cos(radians + (brush.degrees * (Math.PI/180))) * brush.offset);
-            const y = y1 + (Math.sin(radians + (brush.degrees * (Math.PI/180))) * brush.offset);
+            const x = circle.x1 + (Math.cos(radians + (brush.degrees * (Math.PI/180))) * brush.offset);
+            const y = circle.y1 + (Math.sin(radians + (brush.degrees * (Math.PI/180))) * brush.offset);
+            const previousX = circle.previousX1 + (Math.cos(radians + (brush.degrees * (Math.PI/180))) * brush.offset);
+            const previousY = circle.previousY1 + (Math.sin(radians + (brush.degrees * (Math.PI/180))) * brush.offset);
             const color = this.getColor(brush.color);
 
             guides.fillStyle = color;
@@ -138,10 +138,10 @@ export default class Painter {
                 guides.stroke();
             }
 
-            if(brush.link === true && brush.lastPoint !== false) {
+            if(brush.link === true) {
                 canvas.strokeStyle = color;
                 canvas.beginPath();
-                canvas.moveTo(brush.lastPoint.x, brush.lastPoint.y);
+                canvas.moveTo(previousX, previousY);
                 canvas.lineTo(x, y);
                 canvas.stroke();
             } else {
@@ -179,7 +179,7 @@ export default class Painter {
         this.context.clearRect(0,0,this.width, this.height);
     };
 
-    exportPainter (encode = true) {
+    exportPainter (encode = false) {
         const keys = Object.keys(this.settings);
         let data = keys.reduce(function(data, setting) {
             data[setting] = this[setting];
@@ -194,7 +194,7 @@ export default class Painter {
         return data;
     }
 
-    exportBrushes (encode = true) {
+    exportBrushes (encode = false) {
         let data = [];
         this.brushes.forEach(function(brushes, shapeId){
             if(!Array.isArray(brushes)) {
@@ -202,7 +202,7 @@ export default class Painter {
             }
 
             data[shapeId] = brushes.map(brush => {
-                return brush.export(false);
+                return brush.export();
             });
         });
 
@@ -222,15 +222,13 @@ export default class Painter {
 
         offscreenContext.drawImage(this.background,0,0);
         offscreenContext.drawImage(this.canvas,0,0);
-        const image = offscreen.toDataURL("image/png");
-
-        return image;
+        return offscreen.toDataURL("image/png");
     }
 
-    export (encode = true) {
+    export (encode = false) {
         let data = {};
-        const painterData = this.exportPainter(false);
-        const brushesData = this.exportBrushes(false);
+        const painterData = this.exportPainter();
+        const brushesData = this.exportBrushes();
 
         data.painter = painterData;
         data.brushes = brushesData;
@@ -241,12 +239,31 @@ export default class Painter {
 
         return data;
     }
+
+    import (data) {
+        this.importPainter(data.painter);
+        this.importBrushes(data.brushes);
+    }
+
+    importPainter (painterData) {
+        Object.assign(this, painterData);
+    }
+
+    importBrushes (circleBrushesData) {
+        this.brushes = [];
+        circleBrushesData.forEach(function(brushesData, circleId) {
+            if (Array.isArray(brushesData)) {
+                brushesData.forEach(brushOptions => {
+                    const circle = this.engine.list[circleId];
+                    this.addCircleBrush(circle, brushOptions);
+                });
+            }
+        }, this);
+    }
 }
 
 class Brush {
     constructor (painter, options) {
-        this.lastPoint = false;
-
         this.settings = {
             color: (typeof options.color !== 'undefined') ? options.color : painter.color,
             point: (typeof options.point !== 'undefined') ? options.point : painter.point,
@@ -258,7 +275,7 @@ class Brush {
         Object.assign(this, JSON.parse(JSON.stringify(this.settings)));
     }
 
-    export (encode = true) {
+    export (encode = false) {
         const keys = Object.keys(this.settings);
         let data = keys.reduce(function(data, setting) {
             data[setting] = this[setting];