+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0]
+### Added
+
+- Imports painter settings on storage import
+
+### Changed
+
+- Moves control re-rendering to the import function
+- Data is no longer encoded before being stored in local data
+- Default background colour has changes to a blueprint blue 
+- The encode parameter in export functions defaults to false 
+
+### Fixed
+
+- Fixes bug when clearing canvas with the linked brush flag set to true
+
 ## [0.2.0]
 ### Added